compare model runs with vaccination to a baseline run for the branching XPVWRH model where both partial and full vaccination have a given level of efficacy

compare model runs with vaccination to a baseline run for the branching XPVWRH model where both partial and full vaccination have a given level of efficacy

compare_baseline_xpvwrh(
  y,
  baseline,
  uptake_first_dose,
  uptake_second_dose,
  cost_params,
  disc_rate,
  vea,
  vea_p
)

Arguments

y

list of model runs, e.g. created by `run_onevax_xpvwrh`, each list entry refers to a different parameter set

baseline

list of baseline runs e.g. created by `run_onevax_xpvwrh`. Should be same length as `y`

uptake_first_dose

numeric (0-1) either length 1 or same as `y`

uptake_second_dose

numeric (0-1) either length 1 or same as `y`

cost_params

list of cost effectiveness parameters, containing entries `qaly_loss_per_diag_s`, `unit_cost_manage_symptomatic`, `unit_cost_manage_asymptomatic`, `unit_cost_screen_uninfected`. Each entry length 1 or same as `y`

disc_rate

discount rate for cost-effectiveness output per annum, default = 0

vea

= scalar indicating efficacy of full vaccination against aquisition (between 0-1)

vea_p

scalar indicating efficacy of partial vaccination against acquisition (between 0-1)

Value

A list of matrices with rows = time point and columns = parameter set list entries are: `cum_treated` = cumulative number treated in model run `cum_vaccinated` = cumulative vaccinated treated in model run `diag_a` = annual number of asymptomatic diagnoses in model run `diag_s` = annual number of symptomatic diagnoses in model run `treated` = annual number treated in model run `screened` = annual number screened in model run `vaccinated` = annual number vaccinated in model run (includes those getting vbe and boosters) `vbe` = annual number vaccinated before entry in model run `revaccinated` = annual number receiving booster in model run `inc_cum_treated` = cumulative number treated compared to baseline `pv_cases_averted` = cumulative number treated when discounting applied compared to baseline `inc_cum_vaccinated` = cumulative number vaccinated compared to baseline `inc_diag_a` = annual number of asymptomatic diagnoses compared to baseline `inc_diag_s` = annual number of symptomatic diagnoses compared to baseline `inc_treated` = annual number treated compared to baseline `inc_screened` = annual number screened compared to baseline `inc_vaccinated` = annual number vaccinated compared to baseline (includes those getting vbe and boosters) `inc_vbe` = annual number vaccinated before entry compared to baseline `inc_revaccinated` = annual number receiving booster compared to baseline `inc_cum_revaccinated` = cumulative number revaccinated by recieving booster compared to baseline `inc_primary` = annual number receiving primary vaccination (full or partial) compared to baseline (does not include vbe) `inc_cum_primary` = cumulative number of individuals receiving primary vaccination (full or partial) compared to baseline (does not include vbe) `inc_part_to_full`= annual number of individuals who have already received their 1st dose, who receive their 2nd dose compared to baseline `inc_cum_part_to_full` = cumulative number of individuals who have already received their 1st dose, who receive their 2nd dose compared to baseline `inc_doses` = number of doses compared to baseline (per year). Assumes primary vaccination uses 2 doses, booster uses 1 dose. `inc_cum_doses` = cumulative number of doses compared to baseline. Assumes primary vaccination uses 2 doses, booster uses 1 dose. `inc_primary_part_doses` = annual number of 1st doses administered as part of primary vaccination which were not followed by a 2nd dose compared to baseline. `inc_cum_primary_part_doses` = cumulative number of 1st doses administered as part of primary vaccination which were not followed by a 2nd dose compared to baseline. `inc_primary_full_doses` = annual number 1st and 2nd doses administered together as part of primary vaccination compared to baseline. `inc_cum_primary_full_doses` = cumulative number of 1st and 2nd doses administered together as part of primary vaccination compared to baseline. `inc_primary_total_doses` = annual number of primary doses administered compared to baseline. Includes doses contributing to partial and full vaccination `inc_cum_primary_total_doses` = cumulative number of primary doses administered compared to baseline. Includes doses contributing to partial and full vaccination. `inc_booster_doses` = annual number of booster doses administered to waned individuals for revaccination compared to baseline, where 1 dose is needed for full protection. `inc_cum_booster_doses` = cumulative number of booster doses administered to waned individuals for revaccination compared to baseline, where 1 dose is needed for full protection. `inc_vbe_doses` = annual number of vaccinations given before entry, where 2 doses give full protection. `inc_cum_vbe_doses` = cumulative number of vaccinations given before entry, where 2 doses give full protection. `cases_averted_per_dose` = cumulative number of cases (i.e. diagnoses) averted per dose of vaccine `cases_averted_per_dose_pv` = present value of cases_averted_per_dose (i.e. sum of annual numbers discounted at rate `disc_rate` to time 0) `pv_inc_doses` = present value of number of doses compared to baseline (i.e. sum of annual numbers discounted at rate `disc_rate` to time 0) `pv_red_net_cost` = present value of the reduction in net costs compared to baseline (i.e. sum of annual reduction in costs discounted at rate `disc_rate` to time 0) `pv_qaly_gain` = present value of the QALYs gained due to vaccination compared to baseline (i.e. QALYs gained each year discounted at rate `disc_rate` to time 0, and summed) `cet_20k` = cost effectiveness threshold (i.e. £ value per dose at which vaccination becomes cost-effective) calculated using £20,000 / QALY. (pv_red_net_cost + pv_qaly_gain * £20,000) / pv_inc_doses `cet_30k` = cost effectiveness threshold (i.e. £ value per dose at which vaccination becomes cost-effective) calculated using £30,000 / QALY (pv_red_net_cost + pv_qaly_gain * £30,000) / pv_inc_doses `inc_costs_9` = present value of incremental costs assuming £9 / dose. Incremental costs are calculated as: pv_inc_doses * £9 - pv_red_net_cost `inc_costs_18` = present value of incremental costs assuming £18 / dose. Incremental costs are calculated as: pv_inc_doses * £18 - pv_red_net_cost `inc_costs_35` = present value of incremental costs assuming £35 / dose. Incremental costs are calculated as: pv_inc_doses * £35 - pv_red_net_cost `inc_costs_50` = present value of incremental costs assuming £50 / dose. Incremental costs are calculated as: pv_inc_doses * £50 - pv_red_net_cost `inc_costs_70` = present value of incremental costs assuming £70 / dose. Incremental costs are calculated as: pv_inc_doses * £70 - pv_red_net_cost `inc_costs_85` = present value of incremental costs assuming £85 / dose Incremental costs are calculated as: pv_inc_doses * £85 - pv_red_net_cost `vacprotec_full` = number of fully vaccine protected individuals in the population at a given timepoint `vacprotec_part` = number of partially vaccine protected individuals in the population at a given timepoint `vacprotec_total` = total number of vaccine protected individuals in the population (both partially and fully vaccinated) at a given timepoint `vacprotec_full_prop` = proportion of the population experiencing full vaccine protection at a given timepoint `vacprotec_part_prop` = proportion of the population experiencing partial vaccine protection at a given timepoint `vacprotec_total_prop` = total proportion of the population experiencing some form of vaccine protection at a given timepoint `level_vacprotec` = the level of vaccine protection in the population at a given timepoint. Given as the sum of the products of the number of partially and fully vaccinated individuals and the one dose and two dose vaccine efficacies respectively. `inc_diag_a_xwh` = annual number of asymptomatic diagnoses in non- vaccine protected strata compared to an adjusted baseline `inc_diag_s_xwh` = annual number of symptomatic diagnoses in non- vaccine protected strata compared to an adjusted baseline `inc_diag_t_xwh` = annual number of total diagnoses in non-vaccine protected strata compared to an adjusted baseline `inc_diag_a_pvr` = annual number of asymptomatic diagnoses in vaccine protected strata compared to an adjusted baseline `inc_diag_s_pvr` = annual number of symptomatic diagnoses in vaccine protected strata compared to an adjusted baseline `inc_diag_t_pvr` = annual number of total diagnoses in vaccine protected strata compared to an adjusted baseline `inc_cum_diag_a_xwh` = cumulative number of asymptomatic diagnoses in non- vaccine protected strata compared to an adjusted baseline `inc_cum_diag_s_xwh` = cumulative number of symptomatic diagnoses in non- vaccine protected strata compared to an adjusted baseline `inc_cum_diag_t_xwh` = cumulative number of total diagnoses in non-vaccine protected strata compared to an adjusted baseline `inc_cum_diag_a_pvr` = cumulative number of asymptomatic diagnoses in vaccine protected strata compared to an adjusted baseline `inc_cum_diag_s_pvr` = cumulative number of symptomatic diagnoses in vaccine protected strata compared to an adjusted baseline `inc_cum_diag_t_pvr` = cumulative number of total diagnoses in vaccine protected strata compared to an adjusted baseline `percentage_cases_avert` = percentage of cases avoided in each time-point compared to a baseline of no vaccination when cases are counted as successful treatment `cum_percentage_cases_avert`= percentage of cases avoided cumulatively across all timepoints compared to a baseline of no vaccination when cases are counted as successful treatment `percentage_diag_avert` = percentage of cases avoided in each time-point compared to a baseline of no vaccination when cases are counted as the sum of all diagnoses `cum_percentage_diags_avert`= percentage of cases avoided cumulatively across all timepoints compared to a baseline of no vaccination when cases are counted as the sum of all diagnoses `inc_diag_t` = total number of diagnoses in each time point `inc_cum_diag_t` = cumulative total number of diagnoses over time `inc_cum_diag_a` = cumulative number of asymptomatic diagnoses over time `inc_cum_diag_s` = cumulative number of symptomatic diagnoses over time `pv_num_vaccinations` = present value cumulative number of vaccinations