run model from equilibrium with single vaccine at the input efficacy / duration grid locations for n parameter sets

run model from equilibrium with single vaccine at the input efficacy / duration grid locations for n parameter sets

run_grid(
  gono_params,
  init_params,
  cost_params,
  baseline,
  model,
  eff,
  dur,
  vbe = 0,
  strategy = NULL,
  uptake_total = 0,
  uptake_second_dose = uptake_total,
  t_stop = 99,
  full_output = FALSE,
  disc_rate = 0
)

Arguments

gono_params

gono params

init_params

initial state parameters

cost_params

cost effectiveness parameters

baseline

optional input of a baseline to compare to, must be a gonovax_grid object if supplied

model

a gonovax run function, by default `run_onevax`

eff

numeric vector (between 0-1) of vaccine efficacy against acquisition

dur

numeric vector of duration in years of the vaccine

vbe

single numeric indicating (between 0-1), default = 0

strategy,

default is null, no vaccination `VoD`: vaccination on diagnosis, `VoA`: vaccination on attendance, `VoD(L)+VoA(H)`: targeted vaccination (i.e. all diagnosed plus group H on screening)

uptake_total

numeric (0-1) of strategy uptake, default = 0

uptake_second_dose

numeric (0-1) of strategy uptake, default = uptake_total

t_stop

time at which vaccination should stop (years), default = 99

full_output

logical indicating whether full results should be output

disc_rate

discount rate for cost-effectiveness output per annum, default = 0

Value

A `gonovax_grid` object