run model with single vaccine for input parameter sets, either from initialisation or from equilibrium, those with waned vaccines are eligible for revaccination (R), and return to the R stratum

run model with single vaccine for input parameter sets, either from initialisation or from equilibrium, those with waned vaccines are eligible for revaccination (R), and return to the R stratum

run_onevax_xvwrh(
  tt,
  gono_params,
  init_params = NULL,
  dur = 1000,
  vea = 0,
  vei = 0,
  ved = 0,
  ves = 0,
  dur_revax = dur,
  vea_revax = vea,
  vei_revax = vei,
  ved_revax = ved,
  ves_revax = ves,
  vbe = 0,
  primary_uptake = 0,
  booster_uptake = primary_uptake,
  strategy = "VbE",
  t_stop = 99,
  hes = 0
)

Arguments

tt

a numeric vector of times at which the model state is output

gono_params

list of gono params

init_params

= NULL

dur

scalar or numeric vector with same length as `gono_params` giving duration of the vaccine (in years)

vea

scalar or numeric vector with same length as `gono_params` giving efficacy of the vaccine against acquisition (between 0-1)

vei

scalar or numeric vector with same length as `gono_params` giving efficacy of the vaccine against infectiousness (between 0-1)

ved

scalar or numeric vector with same length as `gono_params` giving efficacy of the vaccine against duration (between 0-1)

ves

scalar or numeric vector with same length as `gono_params` giving efficacy of the vaccine against symptoms (between 0-1)

dur_revax

scalar or numeric vector with same length as `gono_params` giving duration of protection for revaccination, default to same as primary

vea_revax

scalar or numeric vector with same length as `gono_params` giving efficacy of revaccination against acquisition, default to same as primary

vei_revax

scalar or numeric vector with same length as `gono_params` giving efficacy of revaccination against infectiousness, default to same as primary

ved_revax

scalar or numeric vector with same length as `gono_params` giving efficacy of revaccination against duration of infection, default to same as primary

ves_revax

scalar or numeric vector with same length as `gono_params` giving efficacy of revaccination against symptoms, default to same as primary

vbe

scalar indicating pc of population vaccinated before entry (between 0-1)

primary_uptake

scalar or numeric vector with same length as 'gono_params' giving proportion of population undertaking primary vaccination as part of strategy

booster_uptake

scalar or numeric vector with same length as 'gono_params' giving proportion of population undertaking booster vaccination after primary vaccination protection has waned. Defaults to supplied value of `primary_uptake`.

strategy

single character string in "VbE", "VoD", "VoD(H)", "VoA", "VoA(H)", "VoD(L)+VoA(H)"

t_stop

time at which vaccination should stop (years)

hes

Proportion of individuals in the population who are vaccine hesitant

Value

A list of transformed model outputs