create vaccination parameters for use in onevax_xpvwrh model

create vaccination parameters for use in onevax_xpvwrh model

vax_params_xpvwrh_trackvt(
  vea = 0,
  vei = 0,
  ved = 0,
  ves = 0,
  vea_revax = vea,
  vei_revax = vei,
  ved_revax = ved,
  ves_revax = ves,
  vea_p = vea,
  vei_p = vei,
  ved_p = ved,
  ves_p = ves,
  dur_va = 1000,
  dur_vb = 1,
  dur_p = dur_va,
  dur_revaxa = dur_va,
  dur_revaxb = 1,
  r1 = 0,
  r2 = 0,
  r1_p = 0,
  r2_p = 0,
  booster_uptake = r1 * r2,
  strategy = NULL,
  vbe = 0,
  t_stop = 99,
  hes = 0,
  n_erlang = 1,
  n_diag_rec = 1,
  years_history = 1
)

Arguments

vea

scalar indicating efficacy of the vaccine against acquisition (between 0-1)

vei

scalar indicating efficacy of the vaccine against infectiousness (between 0-1)

ved

scalar indicating efficacy of the vaccine against duration (between 0-1)

ves

scalar indicating efficacy of the vaccine against symptoms (between 0-1)

vea_revax

scalar indicating efficacy of revaccination against acquisition (between 0-1)

vei_revax

scalar indicating efficacy of revaccination against infectiousness (between 0-1)

ved_revax

scalar indicating efficacy of revaccination against duration of infection (between 0-1)

ves_revax

scalar indicating efficacy of revaccination against symptoms (between 0-1)

vea_p

scalar indicating efficacy of partial vaccination against acquisition (between 0-1)

vei_p

scalar indicating efficacy of partial vaccination against infectiousness (between 0-1)

ved_p

scalar indicating efficacy of partial vaccination against duration (between 0-1)

ves_p

scalar indicating efficacy of partial vaccination against symptoms (between 0-1)

dur_va

duration of time spent in Va stratum after completing a round of primary vaccination (fully vaccinated, accepting first and second dose)

dur_vb

duration of time spent in Vb stratum after completing a round of primary vaccination (fully vaccinated, accepting first and second dose)

dur_p

duration of time spent in the P stratum, partially vaccinated (accepting only the first dose)

dur_revaxa

duration of time spent in Ra stratum after completing a round of primary vaccination (fully vaccinated, accepting first and second dose)

dur_revaxb

duration of time spent in Rb stratum after completing a round of primary vaccination (fully vaccinated, accepting first and second dose)

r1

proportion of population offered vaccine only accepting the first dose, becoming partially vaccinated

r2

proportion of the population who accepted the first dose of the vaccine who go on to accept the second dose, becoming fully vaccinated

r1_p

proportion of partially vaccinated individuals who receive >= one dose when returning to the clinic due to screening or illness

r2_p

proportion of partially vaccinated individuals who receive two doses when returning to the clinic due to screening or illness

booster_uptake

scalar or numeric vector with same length as 'gono_params' giving proportion of population undertaking booster vaccination after primary vaccination protection has waned @param n_diag_rec integer for the number of diagnosis history substrata

strategy

single character string in "VoD", "VoD(H)", "VoA", "VoA(H)", "VoD(L)+VoA(H)". Defaults to NULL i.e. no vaccination

vbe

scalar indicating pc of population vaccinated before entry (between 0-1)

t_stop

time at which vaccination should stop (years)

hes

proportion of population vaccine hesitant

n_erlang

integer giving the number of transitions that need to be made through vaccine-protected strata until that protection has waned

n_diag_rec

number of diagnosis history strata

years_history

number of years that diagnosis history is recorded for

Value

A list parameters in the model input format